requirements.txt
pyworld==0.3.0
librosa==0.8.0
soundfile==0.10.3.post1
tqdm
hydra-core==1.0.0
numpy==1.19.1
kaldiio==2.15.1
